Bugfix to not delete a note with deps

+++ src/features/notes/modals/ModalDeleteNote.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Modal, ModalDialog, Stack, Button, Typography } from "@mui/joy";
+import { Modal, ModalDialog, Stack, Button, Typography, Alert } from "@mui/joy";
 
 import { NoteMetaProps, NoteProps } from "../../../types/notes";
 import { deleteNote } from "../services/deleteNote";
@@ -29,13 +29,30 @@ export const ModalDeleteNote: React.FC<Props> = ({
     handleCloseTab,
 }) => {
     const { accessToken } = useAuth();
-
+    const [errorMessage, setErrorMessage] = useState<string | null>(null);
     const handleDeleteNote = async () => {
-        await deleteNote(myself, currentNote.noteId, accessToken);
-        await deleteData({ storeName: STORES.NOTES, key: currentNote.noteId });
-        setMyNoteMeta(myNoteMeta.filter((note) => note.noteId !== currentNote.noteId));
-        handleCloseTab(currentNote.noteId);
-        setOpenDeleteNote(false);
+        let childExist: boolean;
+        const childNotes = myNoteMeta.filter((note) => note.parentNoteId === currentNote.noteId);
+
+        if (childNotes.length === 0) {
+            childExist = false;
+        } else {
+            childExist = true;
+        }
+
+        if (childExist === false) {
+            // Delete from backend
+            await deleteNote(myself, currentNote.noteId, accessToken);
+            // Delete from indexedDB
+            await deleteData({ storeName: STORES.NOTES, key: currentNote.noteId });
+            // Delete the deleted noteId from the meta object
+            setMyNoteMeta(myNoteMeta.filter((note) => note.noteId !== currentNote.noteId));
+            handleCloseTab(currentNote.noteId);
+            setOpenDeleteNote(false);
+            setErrorMessage(null);
+        } else {
+            setErrorMessage("Can't delete because child note(s) exists.");
+        }
     };
 
     return (
@@ -43,7 +60,10 @@ export const ModalDeleteNote: React.FC<Props> = ({
             <Modal
                 sx={{ zIndex: 10010 }}
                 open={openDeleteNote}
-                onClose={() => setOpenDeleteNote(false)}
+                onClose={() => {
+                    setOpenDeleteNote(false);
+                    setErrorMessage(null);
+                }}
             >
                 <ModalDialog>
                     <Typography level="h4">
@@ -53,12 +73,18 @@ export const ModalDeleteNote: React.FC<Props> = ({
                         </Typography>{" "}
                         ?
                     </Typography>
+                    {errorMessage && errorMessage !== "" && (
+                        <Alert color="danger">{errorMessage}</Alert>
+                    )}
                     <Stack direction="row" spacing={1} sx={{ mt: 2, justifyContent: "center" }}>
                         <Button
                             component="button"
                             color="neutral"
                             variant="outlined"
-                            onClick={() => setOpenDeleteNote(false)}
+                            onClick={() => {
+                                setOpenDeleteNote(false);
+                                setErrorMessage(null);
+                            }}
                         >
                             Cancel
                         </Button>
